pokedex: move constants
