+++ apps/pokedex/src/app/components/constants.ts
